Remove unnecessary third-party components, setup webhook
