+++ .gitignore
@@ -26,6 +26,7 @@ vendor
 logs
 *.log
 *DS_Store
+*.env
 
 # BIN file
 sebatibot
