--- app/config.go
@@ -1,4 +1,0 @@
-package app
-
-// CorsAllowedOrigin is used to define which origins are allowed
-var CorsAllowedOrigin = []string{"*"}
