--- controllers/base.go
@@ -1,13 +1,0 @@
-package controllers
-
-import "net/http"
-
-// BaseController is the common interface for all controllers
-type BaseController interface {
-	Get(w http.ResponseWriter, r *http.Request)
-	GetAll(w http.ResponseWriter, r *http.Request)
-	Post(w http.ResponseWriter, r *http.Request)
-	Put(w http.ResponseWriter, r *http.Request)
-	Delete(w http.ResponseWriter, r *http.Request)
-  PostTelegramUpdate(w http.ResponseWriter, r *http.Request)
-}
