+++ main.go
@@ -1,21 +1,21 @@
 package main
 
 import (
-	// "fmt"
 	"net/http"
-	// "os"
+	"os"
 
-	// "github.com/markbates/going/defaults"
 	"github.com/go-telegram-bot-api/telegram-bot-api"
+	"github.com/gorilla/mux"
 
-	"sebatibot/app"
 	"sebatibot/logger"
 	"github.com/joho/godotenv"
+	"sebatibot/routes"
 )
 
 func main() {
 
-	app := app.App()
+	app := mux.NewRouter()
+	addRoutes(app, routes.UsersRoutes, "/")
 	log := logger.GetLogger()
 
 	err := godotenv.Load()
@@ -23,43 +23,57 @@ func main() {
 		log.Fatal("Error loading .env file")
 	}
 
-	port := defaults.String(os.Getenv("GO_ENV_PORT"), "8000")
-
-	log.Debugf("Starting API server at %s", port)
 	log.Debugf("TOKEN: %s", os.Getenv("TOKEN"))
 
-	//Start the api here
-	// http.ListenAndServe(fmt.Sprintf(":%s", port), app)
 
 	// Initializing Telegram Bot . . .
+	bot, err := tgbotapi.NewBotAPI(os.Getenv("TOKEN"))
+	if err != nil {
+		log.Fatal("Error initializing tgbotapi")
+	}
 
+	bot.Debug = true
+	log.Debugf("Authorized on account %s", bot.Self.UserName)
 
-	// // ---
-	// bot, err := tgbotapi.NewBotAPI("551123175:AAGsVKDZTiQ4Bz4WBHSgP43r7nzATeCAn3U")
-	// if err != nil {
-	// 	log.Panic(err)
-	// }
+  // ngrok for local development
+  // Run ngrok on your local and change webhookUrl by the generated one
+  // Consule README.md for more details
 
-	// bot.Debug = true
+  // webhookUrl := "https://{production_api_endpoint}/" + os.Getenv("TOKEN")
+	webhookUrl := "https://38e44fc0.ngrok.io/" + os.Getenv("TOKEN")
 
-	// log.Debugf("Authorized on account %s", bot.Self.UserName)
+	_, err = bot.SetWebhook(tgbotapi.NewWebhook(webhookUrl))
+	if err != nil {
+		log.Fatal(err)
+	}
 
-	// _, err = bot.SetWebhook(tgbotapi.NewWebhook("https://d073e447.ngrok.io/551123175:AAGsVKDZTiQ4Bz4WBHSgP43r7nzATeCAn3U"))
-	// if err != nil {
-	// 	log.Fatal(err)
-	// }
+	updates := bot.ListenForWebhook("/" + os.Getenv("TOKEN"))
 
-	// updates := bot.ListenForWebhook("/551123175:AAGsVKDZTiQ4Bz4WBHSgP43r7nzATeCAn3U")
-	// // ---
-
-
-	go http.ListenAndServe("127.0.0.1:3000", app)
-	// go http.ListenAndServe("127.0.0.1:8443")
 	log.Debugf("Listening at 127.0.0.1:3000 for Telegram updates. . .")
-
+	go http.ListenAndServe("127.0.0.1:3000", app)
 
 	for update := range updates {
 		log.Debugf("%+v\n", update)
 	}
-	
+}
+
+func addRoutes(app *mux.Router, routes routes.Routes, prefix string) {
+
+	for _, route := range routes {
+
+		if prefix == "" {
+			app.
+				Methods(route.Method).
+				Path(route.Pattern).
+				Name(route.Name).
+				HandlerFunc(route.HandlerFunc)
+		} else {
+			app.
+				PathPrefix(prefix).
+				Methods(route.Method).
+				Path(route.Pattern).
+				Name(route.Name).
+				HandlerFunc(route.HandlerFunc)
+		}
+	}
 }
