--- routes/users_routes.go
@@ -1,31 +1,0 @@
-package routes
-
-import "sebatibot/controllers"
-
-// UsersRoutes contains all urls related to User Resource.
-var UsersRoutes = Routes{
-	Route{
-		Name:        "GetUser",
-		Pattern:     "/{id}",
-		Method:      "GET",
-		HandlerFunc: (&controllers.UsersController{}).Get,
-	},
-	Route{
-		Name:        "GetAllUsers",
-		Pattern:     "",
-		Method:      "GET",
-		HandlerFunc: (&controllers.UsersController{}).GetAll,
-	},
-	Route{
-		Name:        "PostUser",
-		Pattern:     "",
-		Method:      "POST",
-		HandlerFunc: (&controllers.UsersController{}).Post,
-	},
-	Route{
-		Name:        "PostTelegramUpdate",
-		Pattern:     "/551123175:AAGsVKDZTiQ4Bz4WBHSgP43r7nzATeCAn3U",
-		Method:      "POST",
-		HandlerFunc: (&controllers.TelegramController{}).Post,
-	},
-}
